complete!!!!!!!!!!

+++ app/src/main/java/com/example/typing/data/Ranking.kt
@@ -2,5 +2,5 @@ package com.example.typing.data
 
 data class Ranking(
     val uid : String,
-    val typingSpeed: Int
+    val typingSpeed: Double
 )
